VicsekModel

Documentation for VicsekModel.

VicsekModel.VicsekModelParameters

VicsekModel.VicsekModelVariables

VicsekModel.VicsekModelVariables

Base.show

Base.show

VicsekModel.calculate_weights!

VicsekModel.calculate_weights!

VicsekModel.periodic_boundary!

VicsekModel.update!

VicsekModel.update!

VicsekModel.update!

VicsekModel.update_position!

VicsekModel.update_with_anim!

VicsekModel.vicsekmodel

VicsekModel.vicsekmodel_with_anim

VicsekModel.VicsekModelParameters — Type

Fields

L: Length of the square box

ρ: Density of particles

N: Number of particles

r₀: Interaction radius

Δt: Time step

factor: Factor to determine the initial velocity

v₀: Initial velocity

η: Noise strength

alg: Algorithm to update the position and velocity

VicsekModel.VicsekModelVariables — Type

Variables of the Vicsek model

Arguments

pos: Position of particles

vel: Velocity of particles

θ: Angle of particles

S: Complex number to calculate the average angle of neighbors

VicsekModel.VicsekModelVariables — Method

Initialize the variables of the Vicsek model

Arguments

p: Variables of the Vicsek model

seed: keyword variable used as seed for random generator

Base.show — Method

Show the parameters of the Vicsek model

Base.show — Method

Show the variables of the Vicsek model

VicsekModel.calculate_weights! — Method

Calculate the average angle of neighbors

Arguments

pos: Position of particles

θ: Angle of particles

S: Complex number to calculate the average angle of neighbors

r₀: Interaction radius

ix: Index of particles

VicsekModel.calculate_weights! — Method

Calculate the average angle of neighbors with KDTree

Arguments

tree: KDTree to find the neighbors

pos: Position of particles

θ: Angle of particles

S: Complex number to calculate the average angle of neighbors

r₀: Interaction radius

ix: Index of particles

VicsekModel.periodic_boundary! — Method

Periodic boundary condition

Arguments

pos: Position of particles

L: Length of the square box

ix: Index of particles

VicsekModel.update! — Method

Holy trait of update function

VicsekModel.update! — Method

Update the position and velocity of particles using a simple for-loop method

Arguments

var: Variables of the Vicsek model

p: Parameters of the Vicsek model

VicsekModel.update! — Method

Update the position and velocity of particles using KDTree

Arguments

var: Variables of the Vicsek model

p: Parameters of the Vicsek model

VicsekModel.update_position! — Method

Update the position and velocity of particles

Arguments

pos: Position of particles

vel: Velocity of particles

θ: Angle of particles

S: Complex number to calculate the average angle of neighbors

v₀: Initial velocity

η: Noise strength

ix: Index of particles

VicsekModel.update_with_anim! — Method

Update the position and velocity of particles with animation

Arguments

t: Time step

var: Variables of the Vicsek model

p`: Parameters of the Vicsek model

VicsekModel.vicsekmodel — Method

Basic simulation of Vicsek model

Arguments

Tmax: Maximum time step

VicsekModel.vicsekmodel_with_anim — Method

Vicsek model with animation

Arguments

Tmax: Maximum time step